// lib/api/types.ts
export interface IHumanInputFormData {
	form_content: string
	inputs: IHumanInputField[]
	resolved_default_values: Record<string, string>
	user_actions: IHumanInputAction[]
	expiration_time: number
}

export interface IHumanInputField {
	type: 'text_input' | 'select' | 'paragraph' | 'number'
	output_variable_name: string
	default: {
		type: string
		selector: string[]
		value: string
	}
}

export interface IHumanInputAction {
	id: string
	title: string
	button_style: 'primary' | 'default'
}

/**
 * POST /form/human_input/{form_token} 请求体
 */
export interface IHumanInputSubmitBody {
	inputs: Record<string, string>
	action: string
	user: string
}